refactor(#3278): extract emit phases (destructuring/construction/registration) to arrow-phases.ts (slice 2)
Slice 2 lifts the three EMISSION phases of compileArrowAsClosure verbatim into
the sibling module src/codegen/closures/arrow-phases.ts (slice 1 landed the two
analysis/type phases in PR #3082):

- emitClosureParamDestructuring — binding-pattern param destructuring
  (array / tuple-struct / object / externref-host)
- emitClosureConstruction      — construction-site emit (ref.func + capture
  values + TDZ-flag ref cells + struct.new) into the enclosing fctx.body
- registerClosureBindingInfo   — closure-info registration (by-type-idx +
  variable/assignment/module-global closureMap)

closures.ts 3472 -> 2814 LOC; compileArrowAsClosure body 1311 -> 638 (halved) —
all five phase helpers now live in the sibling module. Removed two now-orphaned
imports. Emitted-Wasm byte-identity IDENTICAL 39/39 (gc/standalone/wasi) vs
origin/main. tsc 0, LOC-budget OK, biome clean, tests/issue-3278.test.ts 10/10.
Sets issue #3278 status: done (final slice).

✓
